+++ Common/Server/Infrastructure/Postgres/SchemaMigrations/Index.ts
@@ -151,7 +151,8 @@ import { MigrationName1754315774827 } from "./1754315774827-MigrationName";
 import { MigrationName1754384418632 } from "./1754384418632-MigrationName";
 import { MigrationName1754412708044 } from "./1754412708044-MigrationName";
 import { MigrationName1754415281937 } from "./1754415281937-MigrationName";
-import { UpdateSubscriberNotificationStatusToEnum1754500000000 } from "./1754500000000-UpdateSubscriberNotificationStatusToEnum";
+import { RenameSubscriberNotificationFailedReasonToStatusMessage1754484441976 } from "./1754484441976-RenameSubscriberNotificationFailedReasonToStatusMessage";
+import { UpdateSubscriberNotificationStatusToEnum1754500000000 } from "./1754484441975-UpdateSubscriberNotificationStatusToEnum";
 
 export default [
   InitialMigration,
@@ -307,5 +308,6 @@ export default [
   MigrationName1754384418632,
   UpdateSubscriberNotificationStatusToEnum1754500000000,
   MigrationName1754412708044,
-  MigrationName1754415281937
+  MigrationName1754415281937,
+  RenameSubscriberNotificationFailedReasonToStatusMessage1754484441976
 ];
